+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix for yarn outdated plugin
 - Fix for Capacitor migrate with modern yarn
 - Fix for dialogs of "prepare release" losing focus
+- Angular Schematics for migrating to Signals and inject added
 
 ### Version 1.98